refactor: ApprovalRequirementPolicy を domain/policies/approval/ へ区画化 (#421)
承認要否判定ポリシーと隣接テストを domain/policies/approval/ へ集約。
本体・テストの values 参照を +1 階層調整。消費する見積本体 VO（EstimateType/
Money）への参照は ../../values/ として残し approval/ 外依存を可視化。
参照元は自テストのみ（共移動）。build/eslint(estimate)/test green。
