+++ src/server/subdomains/estimate/domain/policies/approval/ApprovalRequirementPolicy.ts
@@ -1,8 +1,8 @@
 import { ProductCategory } from "@subdomains/product/domain/values/ProductCategory";
-import { ApprovalGoalTier } from "../values/approval/ApprovalGoalTier";
-import { EstimateExemptionReason } from "../values/approval/EstimateExemptionReason";
-import { EstimateType } from "../values/EstimateType";
-import { Money } from "../values/Money";
+import { ApprovalGoalTier } from "../../values/approval/ApprovalGoalTier";
+import { EstimateExemptionReason } from "../../values/approval/EstimateExemptionReason";
+import { EstimateType } from "../../values/EstimateType";
+import { Money } from "../../values/Money";
 
 /**
  * 承認要否の判定結果（§4・ADR-0062）
