+++ src/server/subdomains/estimate/domain/policies/approval/__tests__/ApprovalRequirementPolicy.test.ts
@@ -1,10 +1,10 @@
 import { ProductCategory } from "@subdomains/product/domain/values/ProductCategory";
 import { describe, expect, it } from "vitest";
 import { ApprovalRequirementPolicy } from "../ApprovalRequirementPolicy";
-import { ApprovalGoalTier } from "../../values/approval/ApprovalGoalTier";
-import { EstimateExemptionReason } from "../../values/approval/EstimateExemptionReason";
-import { EstimateType } from "../../values/EstimateType";
-import { Money } from "../../values/Money";
+import { ApprovalGoalTier } from "../../../values/approval/ApprovalGoalTier";
+import { EstimateExemptionReason } from "../../../values/approval/EstimateExemptionReason";
+import { EstimateType } from "../../../values/EstimateType";
+import { Money } from "../../../values/Money";
 
 const yen = (amount: number) => Money.fromMajorUnits(amount);
 const judge = (input: {
